refactor: data 구조 변경
