+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/model/SearchRecipeItem.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.presentation.data.model
+package com.surivalcoding.composerecipeapp.data.model
 
 data class SearchRecipeItem(
     val thumbnailUrl: String,
