+++ app/src/main/java/com/surivalcoding/composerecipeapp/presentation/component/SquareRecipeCard.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.surivalcoding.composerecipeapp.R
-import com.surivalcoding.composerecipeapp.presentation.data.model.RecipeItem
+import com.surivalcoding.composerecipeapp.data.model.RecipeItem
 import com.surivalcoding.composerecipeapp.ui.theme.ComposeRecipeAppTheme
 
 @Composable
